fix Validación Monto Comprobante de Pago

+++ assets/js/ie/CargaDeComprobanteDePago/index.js
@@ -45,7 +45,7 @@ const CargaDeComprobanteDePago = (
   function handleCargarComprobanteDePago(event) {
     event.preventDefault();
 
-    if(parseInt(monto) >= parseInt(gestionPago.montoTotalPorPagar)) {
+    if(parseFloat(monto) >= parseFloat(gestionPago.montoTotalPorPagar)) {
       setMontoError(false);
       event.target.submit()
       return;
